web/requirements.txt
Django==4.2.7
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
websocket-client==1.6.4
redis==5.0.1
python-dotenv==1.0.0
Pillow==10.1.0